+++ torchao/quantization/GPTQ.py
@@ -22,9 +22,11 @@ from torch.utils._pytree import tree_flatten, tree_unflatten
 from .utils import (
     _lm_eval_available,
     _MultiInput,
-    TORCH_VERSION_AFTER_2_3,
+)
+from torchao.utils import (
     find_multiple,
 )
+from torchao.utils import TORCH_VERSION_AFTER_2_3
 from typing import Any, Dict, Optional
 from .unified import Quantizer
