+++ torchao/sparsity/training/__init__.py
@@ -7,7 +7,7 @@ import torch
 from torchao.sparsity.training.autograd import semi_structured_sparsify
 from torchao.sparsity.training.pointwise_ops import CUTLASS_POINTWISE_OP_DISPATCH_TABLE
 
-from torchao.quantization.utils import TORCH_VERSION_AFTER_2_3
+from torchao.utils import TORCH_VERSION_AFTER_2_3
 
 # load pointwise op support, which exists only for CUTLASS
 if TORCH_VERSION_AFTER_2_3:
